Refactor parent task module and controller: rename pTaskModule and pTaskController to parentTaskModule and parentTaskController, update service references, and implement invite handling logic for parent and child invites.
