+++ src/app.module.ts
@@ -6,7 +6,7 @@ import { AdminTaskModule } from './users/admin/task/Atask.module';
 import { AdminModule } from './users/admin/user/Auser.module';
 import { TaskModule } from './users/child/task/task.module';
 import { UsersModule } from './users/child/users/user.module';
-import { pTaskModule } from './users/parent/task/Ptask.module';
+import {  parentTaskModule } from './users/parent/task/Ptask.module';
 import { pUserModule } from './users/parent/user/puser.module';
 import { EventsGateway } from './utils/events/events.gateway';
 import { commonUserModule } from './users/common_user/common_user/common_user.module';
@@ -22,7 +22,7 @@ import { HomeModule } from './home/home.module';
     UsersModule,
     TaskModule,
     pUserModule,
-    pTaskModule,
+    parentTaskModule,
     commonUserModule,
     AdminModule,
     AdminTaskModule,
