+++ src/users/parent/user/puser.service.ts
@@ -1,8 +1,7 @@
 import {
   BadRequestException,
   Injectable,
-  InternalServerErrorException,
-  UnauthorizedException,
+  UnauthorizedException
 } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { JwtService } from '@nestjs/jwt';
@@ -10,34 +9,27 @@ import { InjectModel } from '@nestjs/mongoose';
 import bcrypt from 'bcryptjs';
 import { Model } from 'mongoose';
 import {
-  UserRoleEnum,
-  UserRoleHierarchyEnum,
+  UserRoleEnum
 } from 'src/lib/enums/common.enums';
 import { User } from 'src/Schemas/cSchema/user.schema';
-import { Home } from 'src/Schemas/homeSchema/homeSchema';
-import { Invite } from 'src/Schemas/inviteSchema/inviteSchema';
 import { pUser } from 'src/Schemas/pSchema/pUser.schema';
 import { EmailOptions } from 'src/type';
 import { EventsGateway } from 'src/utils/events/events.gateway';
 import { EmailService } from 'src/utils/services/email';
 import { ParentSignupFieldValidators } from 'src/utils/validators/fieldValidators';
 import {
-  AcceptInviteDto,
   AddChild,
-  ChildInviteDto,
   CreatePatentDto,
   LoginUserDto,
-  ParentInvite,
-  ResendVerificationEmail,
+  ResendVerificationEmail
 } from './dto/Puser.dto';
 
 @Injectable()
 export class pUserService {
   constructor(
     @InjectModel(pUser.name) private pUserModel: Model<pUser>,
-    @InjectModel(Invite.name) private InviteSchema: Model<Invite>,
     @InjectModel(User.name) private userModel: Model<User>,
-    @InjectModel(Home.name) private homeModel: Model<Home>,
+    
     private jwtService: JwtService,
     private emailService: EmailService,
     private eventGateway: EventsGateway,
@@ -273,201 +265,4 @@ export class pUserService {
     }
   }
 
-  async sendParentInvite(parentInvite: ParentInvite) {
-    try {
-      const findUser = await this.InviteSchema.findOne({
-        email: parentInvite.email,
-      });
-      const inviteToken = this.jwtService.sign(
-        { id: parentInvite.email },
-        { secret: process.env.JWT_SECRET, expiresIn: '5M' },
-      );
-      if (findUser) {
-      }
-      const findInvite = await this.InviteSchema.findOne({
-        email: parentInvite.email,
-        homeId: parentInvite.homeId,
-      });
-      if (findInvite) {
-        const updateToken = await this.InviteSchema.findByIdAndUpdate(
-          findInvite._id,
-          {
-            token: inviteToken,
-          },
-        );
-        if (updateToken) {
-          const mailOptions: EmailOptions = {
-            to: parentInvite.email,
-            subject: 'Connect with your homies!!',
-            body: `Hey ${UserRoleHierarchyEnum.CO_LEADER}!! Just accept this invite and ready to connect with your homies  : ${process.env.FRONTEND_DEV_URL}/invite/${updateToken._id}`,
-          };
-          await this.emailService.sendMail(mailOptions);
-          return {
-            message: 'Mail send successfully',
-            status: 200,
-          };
-        }
-      }
-
-      const createInvite = await this.InviteSchema.create({
-        homeId: parentInvite.homeId,
-        email: parentInvite.email,
-        token: inviteToken,
-        roleAssigned: UserRoleHierarchyEnum.CO_LEADER,
-        status: 'pending',
-      });
-
-      if (createInvite) {
-        const mailOptions: EmailOptions = {
-          to: parentInvite.email,
-          subject: 'Connect with your homies!!',
-          body: `Hey ${UserRoleHierarchyEnum.CO_LEADER}!! Just accept this invite and ready to connect with your homies  : ${process.env.FRONTEND_DEV_URL}/invite/${createInvite._id}`,
-        };
-        await this.emailService.sendMail(mailOptions);
-        return {
-          message: 'Mail send successfully',
-          status: 200,
-        };
-      }
-      return {
-        message: 'User not found',
-        status: 404,
-      };
-    } catch (error) {
-      throw new BadRequestException('User not found');
-    }
-  }
-
-  async sendChildrenInvites(childInvitesDto: ChildInviteDto) {
-    const { emails, homeId } = childInvitesDto;
-
-    if (!homeId || !homeId.match(/^[0-9a-fA-F]{24}$/)) {
-      throw new BadRequestException('Invalid homeId format');
-    }
-    const checkHomeExist = await this.homeModel.findById(homeId);
-    if (!checkHomeExist) {
-      throw new BadRequestException('Home not found');
-    }
-
-    const inviteResults = await Promise.allSettled(
-      emails.map(async (email: string) => {
-        const inviteToken = this.jwtService.sign(
-          { email: email },
-          { secret: process.env.JWT_SECRET, expiresIn: '5M' },
-        );
-        const findInvite = await this.InviteSchema.findOne({
-          email: email,
-          homeId: homeId,
-        });
-        if (findInvite) {
-          const updateToken = await this.InviteSchema.findByIdAndUpdate(
-            findInvite._id,
-            {
-              token: inviteToken,
-            },
-          );
-          if (updateToken) {
-            const mailOptions: EmailOptions = {
-              to: email,
-              subject: 'Connect with your homies!!',
-              body: `Hey ${UserRoleHierarchyEnum.MEMBER}!! Just accept this invite and ready to connect with your homies  : ${process.env.FRONTEND_DEV_URL}/invite/${updateToken._id}`,
-            };
-            await this.emailService.sendMail(mailOptions);
-            return {
-              message: 'Mail send successfully',
-              status: 200,
-            };
-          }
-        }
-        const isInviteCreated = await this.InviteSchema.create({
-          homeId: homeId,
-          email: email,
-          roleAssigned: UserRoleHierarchyEnum.MEMBER,
-          status: 'pending',
-          token: inviteToken,
-        });
-        console.log(isInviteCreated, 'isInviteCreated');
-
-        if (isInviteCreated) {
-          const mailOptions: EmailOptions = {
-            to: email,
-            subject: 'Connect with your homies!!',
-            body: `Hey ${UserRoleHierarchyEnum.MEMBER}!! Accept this invite and get ready to connect with your homies: ${process.env.FRONTEND_DEV_URL}/invite/${isInviteCreated._id}`,
-          };
-          await this.emailService.sendMail(mailOptions);
-        }
-      }),
-    );
-
-    const failedInvites = inviteResults.filter(
-      (result) => result.status === 'rejected',
-    );
-
-    if (failedInvites.length > 0) {
-      throw new InternalServerErrorException('Some invites were not created');
-    }
-
-    return {
-      message: 'Invite created successfully',
-      status: 201,
-    };
-  }
-
-  async acceptHomeInvite(invite: { id: string; token: string }) {
-    const { id, token } = invite;
-    const findInvite = await this.InviteSchema.find({ _id: id, token: token });
-
-    if (!findInvite) {
-      throw new BadRequestException('Invalid invite link');
-    }
-
-    return {
-      message: 'Invite accepted successfully',
-      status: 200,
-    };
-  }
-
-  async acceptCoLeaderInvite(acceptInviteDto: AcceptInviteDto) {
-    const findInvite = await this.InviteSchema.findOne({
-      email: acceptInviteDto.email,
-      token: acceptInviteDto.inviteToken,
-    });
-
-    if (!findInvite) {
-      throw new BadRequestException('Invalid invite link');
-    }
-
-    const findUser = await this.pUserModel.findOne({
-      email: findInvite.email,
-    });
-
-    if (!findUser) {
-      throw new BadRequestException(
-        'User not found please create account first',
-      );
-    }
-
-    if (!findUser) {
-      throw new BadRequestException('User not found');
-    }
-
-    const addCoLeader = await this.homeModel.findByIdAndUpdate(
-      findInvite.homeId,
-      {
-        coLeader: findUser._id,
-      },
-    );
-
-    if (!addCoLeader) {
-      throw new BadRequestException('Unable to add co-leader');
-    }
-    if (findInvite) {
-      await this.InviteSchema.findByIdAndDelete(findInvite._id);
-    }
-
-    return {
-      message: 'Co-leader added successfully',
-      status: 200,
-    };
-  }
 }
